requirements.txt
streamlit
openai
pandas
pdfplumber
python-docx
openpyxl
plotly
folium
speechrecognition
langchain
supabase
python-dotenv
